+++ neo/framework/Common.cpp
@@ -193,7 +193,7 @@ private:
 	idStrList					warningList;
 	idStrList					errorList;
 
-	int							gameDLL;
+	uintptr_t					gameDLL;
 
 	idLangDict					languageDict;
